add more redis
