+++ .gitignore
@@ -25,3 +25,4 @@ venv.bak/
 .mypy_cache/
 .dmypy.json
 dmypy.json
+/userdata/
