+++ src/scattermind/system/client/loader.py
@@ -1,5 +1,7 @@
 from typing import Literal, TypedDict
 
+from redipy import RedisConfig
+
 from scattermind.system.client.client import ClientPool
 from scattermind.system.plugins import load_plugin
 
@@ -7,9 +9,13 @@ from scattermind.system.plugins import load_plugin
 LocalClientPoolModule = TypedDict('LocalClientPoolModule', {
     "name": Literal["local"],
 })
+RedisClientPoolModule = TypedDict('RedisClientPoolModule', {
+    "name": Literal["redis"],
+    "cfg": RedisConfig,
+})
 
 
-ClientPoolModule = LocalClientPoolModule
+ClientPoolModule = LocalClientPoolModule | RedisClientPoolModule
 
 
 def load_client_pool(module: ClientPoolModule) -> ClientPool:
@@ -20,4 +26,7 @@ def load_client_pool(module: ClientPoolModule) -> ClientPool:
     if module["name"] == "local":
         from scattermind.system.client.local import LocalClientPool
         return LocalClientPool()
+    if module["name"] == "redis":
+        from scattermind.system.client.redis import RedisClientPool
+        return RedisClientPool(module["cfg"])
     raise ValueError(f"unknown client pool: {module['name']}")
