+++ src/scattermind/system/config/loader.py
@@ -25,6 +25,7 @@ from scattermind.system.readonly.loader import (
     load_readonly_access,
     ReadonlyAccessModule,
 )
+from scattermind.system.redis_util import get_test_config
 
 
 StrategyModule = TypedDict('StrategyModule', {
@@ -75,10 +76,13 @@ def load_config(
 
 def load_test(
         *,
+        is_redis: bool,
         max_store_size: int = 1024 * 1024,
         parallelism: int = 0,
         batch_size: int = 5) -> Config:
     executor_manager: ExecutorManagerModule
+    client_pool: ClientPoolModule
+    data_store: DataStoreModule
     if parallelism > 0:
         executor_manager = {
             "name": "thread",
@@ -91,14 +95,27 @@ def load_test(
             "name": "single",
             "batch_size": batch_size,
         }
-    test_config: ConfigJSON = {
-        "client_pool": {
+    if is_redis:
+        client_pool = {
+            "name": "redis",
+            "cfg": get_test_config(),
+        }
+        data_store = {
+            "name": "redis",
+            "cfg": get_test_config(),
+            "mode": "size",
+        }
+    else:
+        client_pool = {
             "name": "local",
-        },
-        "data_store": {
+        }
+        data_store = {
             "name": "local",
             "max_size": max_store_size,
-        },
+        }
+    test_config: ConfigJSON = {
+        "client_pool": client_pool,
+        "data_store": data_store,
         "executor_manager": executor_manager,
         "queue_pool": {
             "name": "local",
