+++ test/__init__.py
@@ -1,0 +1,4 @@
+import pytest
+
+
+pytest.register_assert_rewrite("test.util")
