+++ test/test_assert.py
@@ -21,9 +21,12 @@ from scattermind.system.torch_util import as_numpy, create_tensor
 
 @pytest.mark.parametrize("batch_size", [1, 5, 11, 20, 50])
 @pytest.mark.parametrize("parallelism", [0, 1, 2, 3])
-def test_assertion_error(batch_size: int, parallelism: int) -> None:
+@pytest.mark.parametrize("is_redis", [False, True])
+def test_assertion_error(
+        batch_size: int, parallelism: int, is_redis: bool) -> None:
     set_debug_output_length(7)
-    config = load_test(batch_size=batch_size, parallelism=parallelism)
+    config = load_test(
+        batch_size=batch_size, parallelism=parallelism, is_redis=is_redis)
     config.load_graph({
         "graphs": [
             {
@@ -64,6 +67,7 @@ def test_assertion_error(batch_size: int, parallelism: int) -> None:
         ],
         "entry": "main",
     })
+    time_start = time.monotonic()
     tasks: list[tuple[TaskId, bool]] = [
         (
             config.enqueue(TaskValueContainer({
@@ -73,7 +77,6 @@ def test_assertion_error(batch_size: int, parallelism: int) -> None:
         )
         for tix in range(20)
     ]
-    time_start = time.monotonic()
     for task_id, _ in tasks:
         assert config.get_status(task_id) == TASK_STATUS_WAIT
     try:
